Updating APIs w/ latest changes from API generator
Moving "extra" stuff into the app layer so we don't have to worry about
overwriting/losing when we re-generate client code for APIs

+++ services/TheBlueAlliance/src/TheBlueAlliance/Model/EventRankingRankingsInner.cs
@@ -39,7 +39,7 @@
             /// <param name="record">record (required).</param>
             /// <param name="sortOrders">Additional year-specific information, may be null. See parent &#x60;sort_order_info&#x60; for details. (required).</param>
             /// <param name="teamKey">The team with this rank. (required).</param>
-        public EventRankingRankingsInner(int dq, Collection<decimal> extraStats, int matchesPlayed, int? qualAverage, int rank, WLTRecord record, Collection<decimal> sortOrders, string teamKey)
+        public EventRankingRankingsInner(int dq, Collection<decimal> extraStats, int matchesPlayed, int? qualAverage, int rank, WLTRecord record, Collection<double> sortOrders, string teamKey)
         {
                       this.Dq = dq;
                       // to ensure "extraStats" is required (not null)
@@ -118,7 +118,7 @@
               /// <value>Additional year-specific information, may be null. See parent &#x60;sort_order_info&#x60; for details.</value>
                 [JsonRequired]
                   [JsonPropertyName("sort_orders")]
-                  public Collection<decimal> SortOrders { get; set; }
+                  public Collection<double> SortOrders { get; set; }
                   
               /// <summary>
               /// The team with this rank.
